Cast to float

+++ src/prediktor/replacement/nlp.py
@@ -34,7 +34,6 @@ def infer_nlp_cache(in_node: SearchNode) -> SearchNode:
     trimmed_ids = _trim_and_pad(input_ids)
     logits, caches = _get_outputs_cache(trimmed_ids, [in_node.cache])
     input_start = trimmed_ids.shape[1] - logits.shape[1]
-    # FIXME: sum in bfloat16 and sum in float differ
     nlp_diff = _nlp_from_logits(input_ids[0][input_start:], logits[0])
     return SearchNode(
         in_node.text, in_node.nlp + nlp_diff,
@@ -59,7 +58,8 @@ def _nlp_from_logits(input_ids: torch.Tensor, logits: torch.Tensor) -> float:
     """
     # prediction for a token is in the previous position
     text_ids = input_ids[1:]
-    softmax = torch.softmax(logits, dim=-1)
+    # casting to avoid different results in bfloat16 and float
+    softmax = torch.softmax(logits.to(float), dim=-1)
     probs = softmax[torch.arange(len(text_ids)), text_ids]
     return -torch.log(probs).sum().item()
